shopfloor: split app/menu to user/menu
