+++ shopfloor/readme/USAGE.rst
@@ -3,4 +3,4 @@ the Demo user. The key to use in the HTTP header ``API-KEY`` is: 72B044F7AC780DA
 
 Curl example::
 
-  curl -X POST "http://localhost:8069/shopfloor/app/menu" -H  "accept: */*" -H  "Content-Type: application/json" -H "API-KEY: 72B044F7AC780DAC"
+  curl -X POST "http://localhost:8069/shopfloor/user/menu" -H  "accept: */*" -H  "Content-Type: application/json" -H "API-KEY: 72B044F7AC780DAC"
