+++ shopfloor/services/__init__.py
@@ -6,6 +6,7 @@ from . import schema_detail
 
 # generic services
 from . import app
+from . import user
 from . import menu
 from . import profile
 from . import scan_anything
