+++ shopfloor/static/description/index.html
@@ -450,7 +450,7 @@ You can change the duration of the retention by changing the System Parameter
 the Demo user. The key to use in the HTTP header <tt class="docutils literal"><span class="pre">API-KEY</span></tt> is: 72B044F7AC780DAC</p>
 <p>Curl example:</p>
 <pre class="literal-block">
-curl -X POST &quot;http://localhost:8069/shopfloor/app/menu&quot; -H  &quot;accept: */*&quot; -H  &quot;Content-Type: application/json&quot; -H &quot;API-KEY: 72B044F7AC780DAC&quot;
+curl -X POST &quot;http://localhost:8069/shopfloor/user/menu&quot; -H  &quot;accept: */*&quot; -H  &quot;Content-Type: application/json&quot; -H &quot;API-KEY: 72B044F7AC780DAC&quot;
 </pre>
 </div>
 <div class="section" id="known-issues-roadmap">
