+++ shopfloor/tests/__init__.py
@@ -1,4 +1,5 @@
 from . import test_app
+from . import test_user
 from . import test_menu
 from . import test_openapi
 from . import test_profile
