+++ shopfloor/tests/test_app.py
@@ -36,52 +36,3 @@ class AppCase(CommonCase):
                 ],
             },
         )
-
-    def test_menu_no_profile(self):
-        """Request /app/menu"""
-        # Simulate the client asking the menu
-        response = self.service.dispatch("menu")
-        menus = self.env["shopfloor.menu"].search([])
-        self.assert_response(
-            response,
-            data={
-                "menus": [
-                    {
-                        "id": menu.id,
-                        "name": menu.name,
-                        "scenario": menu.scenario,
-                        "picking_types": [
-                            {"id": picking_type.id, "name": picking_type.name}
-                            for picking_type in menu.picking_type_ids
-                        ],
-                    }
-                    for menu in menus
-                ]
-            },
-        )
-
-    def test_menu_by_profile(self):
-        """Request /app/menu w/ a specific profile"""
-        # Simulate the client asking the menu
-        menus = self.env["shopfloor.menu"].sudo().search([])
-        menu = menus[0]
-        menu.profile_ids = self.profile
-        (menus - menu).profile_ids = self.profile2
-
-        response = self.service.dispatch("menu")
-        self.assert_response(
-            response,
-            data={
-                "menus": [
-                    {
-                        "id": menu.id,
-                        "name": menu.name,
-                        "scenario": menu.scenario,
-                        "picking_types": [
-                            {"id": picking_type.id, "name": picking_type.name}
-                            for picking_type in menu.picking_type_ids
-                        ],
-                    }
-                ]
-            },
-        )
